gh#758/t11 slice B commits 3-5: delete team_resume.go, redirect team resume, re-enable the slice-A lead-gate skip
Commit 3: deletes the bespoke per-member resume classifier now that
runResumeExec/buildSimpleStartPlan (commit 2) fully replace it.

- internal/cli/team_resume.go, team_resume_test.go: deleted outright
  (2112 + 1409 lines). Every symbol they defined was either genuinely
  dead (the classifier itself, its staged lead-then-dependents tmux
  launch sequencing, its own goal-redelivery machinery) or moved.
- internal/cli/resume_shared.go (new), resume_shared_test.go (new):
  six production helpers and two test helpers that other still-live
  files depend on (canonicalPath, compactNativeValue,
  findMemberRestoreRecord, projectLeadExternalRecordBoundaryViolation,
  parseResumeRoles, teamRoleList; writeMemberLaunchRecord,
  resumeChdir), preserved verbatim with a header explaining why they
  survived the classifier's deletion.
- internal/cli/resume_goal.go: deleted three dead blocks (goal-plan
  construction, retirement-notice delivery, manual-drafting prompt)
  whose sole caller was the deleted classifier; left pre-existing dead
  code (scanResumeGoalRecoveryTransitions and friends, already
  callerless at HEAD before this commit) untouched as out of scope.
- internal/cli/runtime_json.go: deleted the resume JSON envelope
  writers/types that only the classifier produced.
- Test files across the package (resume_test.go, resume_goal_test.go,
  runtime_json_test.go, preflight_test.go, parent_amq_floor_test.go,
  runtime_actions_test.go, liveness_test.go, operator_handoff_card_493_test.go,
  status_replacement_test.go) triaged individually:
  - Deleted outright where the property no longer exists (the staged
    lead-then-dependents launch sequencing is a deliberate, cto-ruled
    behavior change, not a gap; execResumePlan/verifyResumeExecLaunchRecords/
    inspectResumeLeadReady and friends have no live equivalent).
  - Rewritten against the new entry points where the property is still
    real: TestResumeExecRejectsPreFloorAMQBeforeParentMutations and
    TestResumeExecRejectsNonTmuxTerminal now drive runResumeExec
    directly; TestWakeHealthForMemberRejectsPrefixHandleReuse's
    coverage lives on via TestWakeProcessMatcherRejectsForeignRoot.
  - Left already-skipped where cto's ruling defers the equivalent to
    slice C (goal-blocked-recovery, reorient/reattach, the cross-command
    status/resume liveness-agreement invariant) -- only the dead-symbol
    bodies were stripped so the package compiles; the skip rationale is
    unchanged.
  - TestResumeExecPrintsOperatorHandoffCard stays skipped, but for a new
    reason found while triaging it: runSimpleStartWithRequest never
    calls printOperatorHandoffCard at all (only team_launch.go's legacy
    executeTeamLaunch path does), so resume --exec silently regressed
    the #493 fix it was written to protect. Filed as gh#788, not fixed
    here (orthogonal to this commit's scope, same as gh#787).

Commit 4: `team resume` is now a hard-error redirect (runTeamResumeDeleted,
team.go) rather than a deprecation notice, per gh#758's "deleted, not
deprecated" ruling -- it always errors, pointing at `amq-squad resume`
(`--exec` for the same launch behavior).

Commit 5: TestResumeDoesNotRestoreDeadExternalLeadRecord (team_lead_test.go)
re-enabled against the real mechanism commit 2 built for it:
resolveResumeLeadGate's external_lead_record_dead required action
(resume_required_actions.go). The underlying safety fact is unchanged --
projectLeadExternalRecordBoundaryViolation (resume_shared.go) still flags a
dead externally-adopted lead record -- but it now surfaces as a hard
refusal on resume --exec instead of a blocked row in the plan-only preview,
which no longer reads launch.Record at all.

Verified: go build ./..., go vet ./internal/cli/..., and the full
go test ./internal/cli/... pass with zero regressions. The only
failure (TestRealPTYWaitPostureRefusesNamedGateBeforeTimeout) is a
known pre-existing, environment-dependent flake unrelated to this
change (last touched in the v2.28 era; fails identically in isolation
against a fake amq binary that doesn't support `shell-setup`).
