+++ internal/cli/resume_goal.go
@@ -1,15 +1,12 @@
 package cli
 
 import (
-	"bufio"
 	"crypto/sha256"
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"time"
 	"unicode"
@@ -162,253 +159,6 @@ type resumeGoalSendSnapshot struct {
 	TeamModTime   int64
 	LaunchDigest  string
 	LaunchModTime int64
-}
-
-// resumeNativeGoalRecovery is read-only evidence that a restored member
-// recorded a native goal which Codex has blocked. Resume never injects the
-// recovery command: the operator must inspect the exact pane before entering
-// /goal resume.
-type resumeNativeGoalRecovery struct {
-	Role     string `json:"role"`
-	Handle   string `json:"handle,omitempty"`
-	Action   string `json:"action"`
-	Detail   string `json:"detail,omitempty"`
-	Guidance string `json:"guidance"`
-}
-
-const nativeGoalBlockedResumeGuidance = "Inspect the exact recovered pane for this profile and session, then enter /goal resume manually. Do not automatically redeliver the saved goal."
-
-func resumeNativeGoalBlockedRecoveries(plans []resumePlan) []resumeNativeGoalRecovery {
-	var recoveries []resumeNativeGoalRecovery
-	for _, plan := range plans {
-		if plan.RestoreRecord == nil || !nativeGoalBindingBlocked(plan.RestoreRecord.GoalBinding) {
-			continue
-		}
-		detail := boundedResumeDisplay(plan.RestoreRecord.GoalBinding.Detail, 240)
-		if detail == "" {
-			detail = "native /goal binding is blocked"
-		}
-		recoveries = append(recoveries, resumeNativeGoalRecovery{
-			Role:     plan.Role,
-			Handle:   plan.Handle,
-			Action:   string(plan.Action),
-			Detail:   detail,
-			Guidance: nativeGoalBlockedResumeGuidance,
-		})
-	}
-	return recoveries
-}
-
-func writeResumeNativeGoalBlockedRecoveries(out io.Writer, recoveries []resumeNativeGoalRecovery) {
-	if len(recoveries) == 0 {
-		return
-	}
-	fmt.Fprintln(out, "# Native goal recovery required")
-	for _, recovery := range recoveries {
-		identity := recovery.Role
-		if recovery.Handle != "" && recovery.Handle != recovery.Role {
-			identity += " (handle " + recovery.Handle + ")"
-		}
-		fmt.Fprintf(out, "# %s (%s): %s\n", identity, recovery.Action, recovery.Detail)
-		fmt.Fprintf(out, "# Recovery: %s\n", recovery.Guidance)
-	}
-	fmt.Fprintln(out)
-}
-
-// buildResumeGoalPlan computes read-only evidence from the exact per-member
-// restore selection. It never claims or creates an attempt.
-func buildResumeGoalPlan(t team.Team, profile, workstream string, plans []resumePlan, force, noBootstrap bool) runwizard.ResumeGoalPlan {
-	result := runwizard.ResumeGoalPlan{SchemaVersion: resumeGoalPlanSchemaVersion, Action: "unavailable"}
-	lead := strings.TrimSpace(t.Lead)
-	if lead == "" && len(t.Members) == 1 {
-		lead = strings.TrimSpace(t.Members[0].Role)
-	}
-	result.LeadRole = lead
-	finish := func(reason string) runwizard.ResumeGoalPlan {
-		result.Eligible = false
-		result.Reason = reason
-		result.EvidenceDigest = resumeGoalEvidenceDigest(result)
-		return result
-	}
-	finishAs := func(action, reason string) runwizard.ResumeGoalPlan {
-		result.Action = action
-		return finish(reason)
-	}
-	if lead == "" {
-		return finish("team has no configured or inferable lead")
-	}
-	member, ok := teamMemberByRole(t, lead)
-	if !ok {
-		return finish("configured lead is not a current team member")
-	}
-	var selected *resumePlan
-	for i := range plans {
-		if plans[i].Role == lead {
-			selected = &plans[i]
-			break
-		}
-	}
-	if selected == nil {
-		return finish("lead is not selected by this resume plan")
-	}
-	result.LeadHandle = selected.Handle
-	result.LeadResumeAction = string(selected.Action)
-	if force {
-		return finish("forced duplicate resume cannot redeliver a saved goal")
-	}
-	if selected.Action != resumeRestore {
-		return finish("lead is not a fresh re-orient restore")
-	}
-	if selected.RestoreRecord == nil {
-		return finish("lead restore record is unavailable")
-	}
-	rec := *selected.RestoreRecord
-	result.SavedConversation = rec.Conversation != ""
-	if canonicalPath(rec.TeamHome) != canonicalPath(t.Project) {
-		return finish("lead restore record does not exactly match the configured team home")
-	}
-	if canonicalPath(rec.CWD) == "" || canonicalPath(rec.CWD) != canonicalPath(member.EffectiveCWD(t.Project)) {
-		return finish("lead restore record does not exactly match the configured cwd")
-	}
-	if rec.Role != lead || rec.Handle != selected.Handle ||
-		strings.TrimSpace(rec.Session) != strings.TrimSpace(workstream) || !squadnamespace.ProfilesEqual(rec.TeamProfile, profile) ||
-		rec.Binary != member.Binary {
-		return finish("lead restore record identity does not match the selected team member")
-	}
-	ns := squadnamespace.Resolve(t.Project, profile, workstream)
-	if canonicalPath(rec.Root) != canonicalPath(ns.AMQRoot) {
-		return finish("lead restore record does not exactly match the workstream root")
-	}
-	if rec.Tmux != nil && rec.Tmux.Target == "adopted" {
-		return finish("adopted lead pane is not eligible for saved-goal redelivery")
-	}
-	if rec.GoalBinding == nil {
-		return finish("lead restore record has no saved goal binding")
-	}
-	result.Action = "blocked"
-	binding := *rec.GoalBinding
-	result.BindingMode = binding.Mode
-	result.BindingNative = binding.NativeGoal
-	result.BindingSource = binding.Source
-	result.BindingDigest = digestJSON(binding)
-	result.BindingCommandDigest = digestBytes([]byte(binding.Command))
-	contract, err := goalDeliveryContractForBinary(member.Binary)
-	if err != nil {
-		return finish("saved goal binding uses an unsupported lead binary")
-	}
-	if binding.Mode != contract.Mode || binding.NativeGoal != contract.NativeGoal || binding.Source != "goal-control" {
-		return finish("saved goal binding is not an original goal-control delivery")
-	}
-	goal, attemptID, err := goalBindingPayload(&binding, contract)
-	if err != nil {
-		return finish("saved goal binding is invalid: " + err.Error())
-	}
-	if expected := contract.prompt(goal, t, profile, workstream, lead, attemptID); !exactGoalBinding(&binding, contract, goal, attemptID, expected, "goal-control") {
-		return finish("saved goal binding does not exactly match the generated goal-control command")
-	}
-	result.Goal = goal
-	result.OriginalAttemptID = attemptID
-	if result.SavedConversation {
-		return finishAs("skip", "lead will reattach its saved conversation and keeps the saved goal in context")
-	}
-	if rec.External {
-		return finish("lead restore record is external and not managed by resume")
-	}
-	if noBootstrap {
-		return finish("resume explicitly disables bootstrap re-orientation")
-	}
-	if rec.BootstrapExpectation == nil || !rec.BootstrapExpectation.Required {
-		return finish("saved lead launch did not have bootstrap re-orientation enabled")
-	}
-	attemptPath, err := goalAttemptPath(t.Project, profile, workstream, attemptID)
-	if err != nil {
-		return finish("saved goal attempt id is invalid")
-	}
-	attemptBytes, err := os.ReadFile(attemptPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			result.AttemptState = "missing"
-			return finish("original goal attempt record is missing")
-		}
-		result.AttemptState = "unreadable"
-		return finish("original goal attempt record is unreadable")
-	}
-	result.AttemptDigest = digestBytes(attemptBytes)
-	var attempt goalAttemptRecord
-	if err := json.Unmarshal(attemptBytes, &attempt); err != nil {
-		result.AttemptState = "invalid"
-		return finish("original goal attempt record is corrupt")
-	}
-	if err := validateResumeGoalAttempt(attempt, t.Project, profile, workstream, lead, selected.Handle, goal, attemptID, ns); err != nil {
-		result.AttemptState = "mismatched"
-		return finish("original goal attempt record is mismatched: " + err.Error())
-	}
-	result.AttemptState = "recorded"
-	claimPath := goalAttemptClaimPath(attemptPath)
-	claimBytes, err := os.ReadFile(claimPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			result.ClaimState = "unclaimed"
-			return finish("original goal attempt is not settled: claim is missing")
-		}
-		result.ClaimState = "unreadable"
-		return finish("original goal claim is unreadable")
-	}
-	result.ClaimDigest = digestBytes(claimBytes)
-	var claim goalAttemptClaim
-	if err := json.Unmarshal(claimBytes, &claim); err != nil {
-		result.ClaimState = "invalid"
-		return finish("original goal claim is corrupt")
-	}
-	if err := validateResumeGoalClaim(claim, attempt); err != nil {
-		result.ClaimState = "mismatched"
-		return finish("original goal claim is mismatched: " + err.Error())
-	}
-	result.ClaimState = "claimed"
-	result.ClaimRoute = claim.Route
-	result.TransitionID = resumeGoalTransitionID(result.OriginalAttemptID, result.BindingDigest)
-	transitionPath, err := resumeGoalTransitionPath(t.Project, profile, workstream, result.TransitionID)
-	if err != nil {
-		result.TransitionState = "invalid"
-		return finish("goal redelivery transition identity is invalid")
-	}
-	if transitionBytes, readErr := os.ReadFile(transitionPath); readErr == nil {
-		result.TransitionDigest = digestBytes(transitionBytes)
-		var transition resumeGoalTransitionRecord
-		if json.Unmarshal(transitionBytes, &transition) != nil || validateResumeGoalTransitionPlan(transition, t.Project, profile, workstream, result) != nil {
-			result.TransitionState = "mismatched"
-			return finish("a mismatched durable goal redelivery transition already exists")
-		}
-		result.RecoveryAttemptID = transition.NewAttemptID
-		result.TransitionState = "reserved"
-		consumedPath := resumeGoalTransitionConsumedPath(transitionPath)
-		if consumedBytes, consumedErr := os.ReadFile(consumedPath); consumedErr == nil {
-			var consumed resumeGoalTransitionConsumed
-			if json.Unmarshal(consumedBytes, &consumed) != nil || consumed.SchemaVersion != resumeGoalTransitionSchemaVersion || consumed.TransitionID != transition.TransitionID || consumed.NewAttemptID != transition.NewAttemptID || consumed.ConsumedAt.IsZero() {
-				result.TransitionState = "mismatched"
-				return finish("durable goal redelivery transition completion is mismatched")
-			}
-			result.TransitionState = "consumed"
-			result.Action = "retry"
-			result.RecoveryCommand = goalManualDeliveryCommand(t.Project, profile, workstream, result.Goal)
-			return finish("a consumed durable goal redelivery transition may have reached the pane; manually retry only its exact claim-once attempt")
-		} else if !os.IsNotExist(consumedErr) {
-			result.TransitionState = "unreadable"
-			return finish("goal redelivery transition completion evidence is unreadable")
-		}
-		result.Action = "continue"
-		result.RecoveryCommand = goalManualDeliveryCommand(t.Project, profile, workstream, result.Goal)
-		return finish("a durable goal redelivery transition is reserved; manually continue its exact claim-once attempt")
-	} else if !os.IsNotExist(readErr) {
-		result.TransitionState = "unreadable"
-		return finish("goal redelivery transition evidence is unreadable")
-	}
-	result.TransitionState = "absent"
-	result.Action = "redeliver"
-	result.Eligible = true
-	result.Reason = "settled original goal can be redelivered as a new claim-once attempt after lead re-orientation"
-	result.EvidenceDigest = resumeGoalEvidenceDigest(result)
-	return result
 }
 
 func resumeGoalTransitionID(attemptID, bindingDigest string) string {
@@ -664,114 +414,6 @@ func cloneGoalBinding(binding *launch.GoalBinding) *launch.GoalBinding {
 	}
 	copy := *binding
 	return &copy
-}
-
-func writeResumeGoalPlan(out io.Writer, plan runwizard.ResumeGoalPlan) {
-	if plan.SchemaVersion == 0 {
-		return
-	}
-	fmt.Fprintln(out)
-	fmt.Fprintln(out, "# recorded goal redelivery")
-	fmt.Fprintf(out, "# lead: %s (%s)\n", plan.LeadRole, plan.LeadHandle)
-	fmt.Fprintf(out, "# eligible: %t\n", plan.Eligible)
-	fmt.Fprintf(out, "# selected: %t\n", plan.Selected)
-	fmt.Fprintf(out, "# action: %s\n", plan.Action)
-	fmt.Fprintf(out, "# reason: %s\n", boundedResumeDisplay(plan.Reason, 240))
-	if plan.Goal != "" {
-		fmt.Fprintf(out, "Recorded goal: %s\n", boundedResumeDisplay(plan.Goal, 240))
-	}
-	if plan.Eligible {
-		fmt.Fprintln(out, "goal redelivery via pane injection is retired (gh#761); after the lead is verified live, deliver manually with `amq-squad goal --goal TEXT`.")
-	}
-	if plan.RecoveryCommand != "" {
-		fmt.Fprintf(out, "Recovery evidence for exact prior attempt %s (manual; retired pane-injection redelivery replaced by an ordinary goal send):\n  %s\n", plan.RecoveryAttemptID, plan.RecoveryCommand)
-	}
-}
-
-func promptResumeGoalRedelivery(in io.Reader, out io.Writer, plan runwizard.ResumeGoalPlan) (bool, error) {
-	if in == nil {
-		in = os.Stdin
-	}
-	if out == nil {
-		out = os.Stderr
-	}
-	fmt.Fprintf(out, "Recorded goal for %s: %s\nRedeliver it as a new claim-once attempt after the lead is verified live? [y/N] ", plan.LeadRole, boundedResumeDisplay(plan.Goal, 240))
-	line, err := readWizardLine(in)
-	if err != nil && err != io.EOF {
-		return false, err
-	}
-	answer := strings.ToLower(strings.TrimSpace(line))
-	return answer == "y" || answer == "yes", nil
-}
-
-func readWizardLine(in io.Reader) (string, error) {
-	if reader, ok := in.(*bufio.Reader); ok {
-		line, err := reader.ReadString('\n')
-		if err == io.EOF && line != "" {
-			return line, nil
-		}
-		return line, err
-	}
-	line, err := bufio.NewReader(in).ReadString('\n')
-	if err == io.EOF && line != "" {
-		return line, nil
-	}
-	return line, err
-}
-
-func boundedResumeDisplay(value string, maxRunes int) string {
-	quoted := []rune(strconv.QuoteToGraphic(value))
-	if maxRunes > 1 && len(quoted) > maxRunes {
-		quoted = append(quoted[:maxRunes-1], '…')
-	}
-	return string(quoted)
-}
-
-// deliverResumeGoalAfterLaunch used to redeliver a goal to a freshly
-// relaunched lead via runtime pane injection (`goal start
-// --resume-transition ...`, reserving a durable transition first). gh#761
-// deleted that entire delivery mechanism -- goal delivery is now launch-time
-// InitialInput, compiled by launchintent from the brief/goal, with no
-// runtime redelivery equivalent yet. Automatic redelivery is retired for
-// ALL sessions in v2.31.0, not conditioned on launch backend: the
-// `--redeliver-goal` flag on `resume --exec` stays registered (existing
-// invocations do not become flag-parse errors) but this function is now a
-// deprecation redirect -- it prints the manual replacement command and
-// returns nil so `resume --exec` itself never fails just because the
-// retired flag was passed. t11 (gh#758's ResumePolicy/OnLive rebuild) is
-// where a relaunch carries the goal at launch time via GoalPrompt, the
-// correct replacement for pane redelivery; see task/t11 for the inherited
-// acceptance note.
-func deliverResumeGoalAfterLaunch(t team.Team, profile, workstream string, results []resumeExecLaunchResult, plan runwizard.ResumeGoalPlan) error {
-	if !plan.Eligible || !plan.Selected {
-		return usageErrorf("resume goal redelivery requires an eligible selected goal plan")
-	}
-	var result *resumeExecLaunchResult
-	for i := range results {
-		if results[i].Check.Role == plan.LeadRole {
-			result = &results[i]
-			break
-		}
-	}
-	if result == nil {
-		return &PartialError{Message: "resume launched members but could not identify the selected lead for goal redelivery"}
-	}
-	quietNotice("goal redelivery via pane injection retired in v2.31.0; deliver with: %s\n", goalManualDeliveryCommand(t.Project, profile, workstream, plan.Goal))
-	return nil
-}
-
-// goalManualDeliveryCommand builds the complete, literally-runnable
-// `amq-squad goal --goal TEXT` replacement command gh#761 prints wherever
-// this package used to either automatically redeliver a goal or print a
-// manual recovery command for a now-deleted subcommand (goal
-// start/retry-attempt). Every value is shell-quoted; nothing is elided.
-func goalManualDeliveryCommand(project, profile, session, goal string) string {
-	args := []string{"amq-squad", "goal", "--goal", goal, "--project", project, "--profile", profile, "--session", session}
-	quoted := make([]string, 0, len(args))
-	for _, arg := range args {
-		quoted = append(quoted, shellQuote(arg))
-	}
-	return strings.Join(quoted, " ")
 }
 
 // resumeGoalRecoveryScanOptions carries the delivery identity used to decide
