+++ internal/cli/resume_test.go
@@ -4,15 +4,12 @@ import (
 	"encoding/json"
 	"os"
 	"path/filepath"
-	"reflect"
 	"strings"
 	"testing"
 	"time"
 
-	"github.com/omriariav/amq-squad/v2/internal/bootstrapack"
 	"github.com/omriariav/amq-squad/v2/internal/launch"
 	"github.com/omriariav/amq-squad/v2/internal/team"
-	"github.com/omriariav/amq-squad/v2/internal/tmuxpane"
 )
 
 // TestResumeAliasPlanIsByteIdenticalToPlan is gh#758's named acceptance
@@ -298,93 +295,11 @@ func TestRunResumeReorientsSeatWithoutConversation(t *testing.T) {
 }
 
 func TestRunResumeSurfacesNativeGoalBlockedRecovery(t *testing.T) {
-	t.Skip("gh#758/t11: deferred to slice C, not deleted or rewritten -- this is exactly the goal-blocked-recovery dormancy already flagged on task/t11 (inherited from t9/gh#761): rec.GoalBinding is never set for a launchapi-launched lead, and the plan-only path no longer reads launch.Record at all regardless. cto's ruling: build an equivalent mechanism on the new resume architecture once t9's real InitialInput shape is available (merged, main 3197386), not before.")
-	dir := t.TempDir()
-	base := setupFakeAMQSessionRoots(t)
-	resumeChdir(t, dir)
-	if err := team.Write(dir, team.Team{
-		Workstream:    "issue-447",
-		Orchestrated:  true,
-		Lead:          "cto",
-		ExecutionMode: executionModeProjectLead,
-		Members:       []team.Member{{Role: "cto", Binary: "codex", Handle: "cto", Session: "issue-447"}},
-	}); err != nil {
-		t.Fatal(err)
-	}
-	writeMemberLaunchRecord(t, base, "issue-447", "cto", launch.Record{
-		CWD: dir, Binary: "codex", Role: "cto", StartedAt: time.Now(),
-		GoalBinding: &launch.GoalBinding{
-			Mode:       "native_goal_blocked",
-			NativeGoal: true,
-			Source:     "goal-runtime",
-			Command:    `/goal --goal "ship"`,
-			Detail:     "Goal blocked (/goal resume)",
-		},
-	})
-
-	plain, _, err := captureOutput(t, func() error { return runResume(nil) })
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !strings.Contains(plain, "Native goal recovery required") || !strings.Contains(plain, "then enter /goal resume manually") || !strings.Contains(plain, "Do not automatically redeliver") {
-		t.Fatalf("resume plan did not surface safe blocked-goal recovery:\n%s", plain)
-	}
-
-	jsonOut, _, err := captureOutput(t, func() error { return runResume([]string{"--json"}) })
-	if err != nil {
-		t.Fatal(err)
-	}
-	env := decodeJSONEnvelope[resumeEnvelopeData](t, jsonOut)
-	if len(env.Data.NativeGoalBlockedRecovery) != 1 || env.Data.NativeGoalBlockedRecovery[0].Role != "cto" || !strings.Contains(env.Data.NativeGoalBlockedRecovery[0].Guidance, "/goal resume") {
-		t.Fatalf("resume JSON did not surface blocked-goal recovery:\n%s", jsonOut)
-	}
+	t.Skip("gh#758/t11: deferred to slice C, not deleted or rewritten -- rec.GoalBinding is never set for a launchapi-launched lead, and the plan-only path no longer reads launch.Record at all regardless. cto's ruling: build an equivalent mechanism on the new resume architecture once t9's real InitialInput shape is available (merged, main 3197386), not before. Body stripped in commit 3: it decoded via resumeEnvelopeData, a runtime_json.go type deleted along with team_resume.go.")
 }
 
 func TestRunResumeExecSurfacesBlockedGoalRecoveryForMixedRoster(t *testing.T) {
-	t.Skip("gh#758/t11: deferred to slice C, not deleted or rewritten -- this is --exec's own copy of the same goal-blocked-recovery dormancy already flagged on task/t11 for the plan-only path (TestRunResumeSurfacesNativeGoalBlockedRecovery, resume_test.go). --exec now drives runResumeExec/simple_start's shared machinery (commit 2), which does not read launch.Record.GoalBinding or emit '# Recovery:' guidance at all. cto's ruling: build an equivalent mechanism on the new resume architecture once t9's real InitialInput shape is available (merged, main 3197386), not before -- same as the plan-only path's deferral.")
-	dir := t.TempDir()
-	base := setupFakeAMQSessionRoots(t)
-	resumeChdir(t, dir)
-	if err := team.Write(dir, team.Team{
-		Workstream: "issue-447", Orchestrated: true, Lead: "cto", ExecutionMode: executionModeProjectLead,
-		Members: []team.Member{
-			{Role: "cto", Binary: "codex", Handle: "cto", Session: "issue-447"},
-			{Role: "qa", Binary: "codex", Handle: "qa", Session: "issue-447"},
-			{Role: "fullstack", Binary: "codex", Handle: "fullstack", Session: "issue-447"},
-		},
-	}); err != nil {
-		t.Fatal(err)
-	}
-	for _, row := range []struct {
-		role    string
-		binding *launch.GoalBinding
-	}{
-		{role: "cto", binding: &launch.GoalBinding{Mode: "native_goal_blocked", NativeGoal: true, Detail: "lead blocked"}},
-		{role: "qa", binding: &launch.GoalBinding{Mode: "native_goal_blocked", NativeGoal: true, Detail: "worker blocked"}},
-		{role: "fullstack", binding: &launch.GoalBinding{Mode: "native_goal", NativeGoal: true, Detail: "delivered"}},
-	} {
-		writeMemberLaunchRecord(t, base, "issue-447", row.role, launch.Record{CWD: dir, Binary: "codex", Role: row.role, Handle: row.role, Session: "issue-447", StartedAt: time.Now(), GoalBinding: row.binding})
-	}
-	oldRun, oldVerify, oldReady := runTmuxLaunchPlanForResume, verifyResumeExecLaunchRecordsNow, verifyResumeLeadReadyNow
-	runTmuxLaunchPlanForResume = func(tmuxLaunchPlan) error { return nil }
-	verifyResumeLeadReadyNow = func(resumeExecLaunchCheck) error { return nil }
-	verifyResumeExecLaunchRecordsNow = func(checks []resumeExecLaunchCheck, _ map[string]resumeExecLaunchSnapshot) []resumeExecLaunchResult {
-		out := make([]resumeExecLaunchResult, 0, len(checks))
-		for _, check := range checks {
-			out = append(out, resumeExecLaunchResult{Check: check, State: resumeExecLaunchStateLaunched})
-		}
-		return out
-	}
-	t.Cleanup(func() {
-		runTmuxLaunchPlanForResume, verifyResumeExecLaunchRecordsNow, verifyResumeLeadReadyNow = oldRun, oldVerify, oldReady
-	})
-	_, stderr, err := captureOutput(t, func() error { return runResume([]string{"--exec", "--stagger", "0"}) })
-	if err != nil {
-		t.Fatalf("resume --exec: %v\nstderr:\n%s", err, stderr)
-	}
-	if strings.Count(stderr, "# Recovery:") != 2 || !strings.Contains(stderr, "cto") || !strings.Contains(stderr, "qa") || strings.Contains(stderr, "fullstack") || !strings.Contains(stderr, "then enter /goal resume manually") {
-		t.Fatalf("mixed exec recovery output = %q", stderr)
-	}
+	t.Skip("gh#758/t11: deferred to slice C, not deleted or rewritten -- this is --exec's own copy of the same goal-blocked-recovery dormancy already flagged on task/t11 for the plan-only path (TestRunResumeSurfacesNativeGoalBlockedRecovery, resume_test.go). --exec now drives runResumeExec/simple_start's shared machinery (commit 2), which does not read launch.Record.GoalBinding or emit '# Recovery:' guidance at all. cto's ruling: build an equivalent mechanism on the new resume architecture once t9's real InitialInput shape is available (merged, main 3197386), not before -- same as the plan-only path's deferral. Body stripped in commit 3: its stubs (runTmuxLaunchPlanForResume/verifyResumeExecLaunchRecordsNow/verifyResumeLeadReadyNow) were team_resume.go internals, now deleted.")
 }
 
 func TestRunResumeRejectsFreshFlag(t *testing.T) {
@@ -617,495 +532,32 @@ func TestRunResumeRejectsExecWithDryRun(t *testing.T) {
 	}
 }
 
-// TestExecResumePlanRefusesBlockedMembersUnlessForced covers the contract
-// that resume --exec is not a backdoor around live-agent protection: any
-// member in action=blocked aborts the run unless --force-duplicate.
-func TestExecResumePlanRefusesBlockedMembersUnlessForced(t *testing.T) {
-	t.Run("blocked aborts without force", func(t *testing.T) {
-		err := execResumePlan(
-			team.Team{Project: t.TempDir(), Members: []team.Member{{Role: "cto"}}},
-			team.DefaultProfile,
-			"issue-96",
-			[]resumePlan{
-				{Role: "cto", Action: resumeBlocked, Note: "wake+presence", Command: ""},
-			},
-			resumeExecOptions{Enabled: true, Terminal: "tmux", Target: "current-window", Layout: "vertical"},
-			false,
-		)
-		if err == nil || !strings.Contains(err.Error(), "blocked") {
-			t.Fatalf("blocked member should abort: %v", err)
-		}
-		if !strings.Contains(err.Error(), "--force-duplicate") {
-			t.Errorf("error should mention escape hatch: %v", err)
-		}
-	})
-}
-
-// TestExecResumePlanNothingToLaunch covers the all-live scenario: every
-// member is already running, so there is nothing to send through the
-// terminal backend. Exit cleanly with a notice rather than opening an
-// empty pane.
-func TestExecResumePlanNothingToLaunch(t *testing.T) {
-	dir := t.TempDir()
-	stdout, _, err := captureOutput(t, func() error {
-		return execResumePlan(
-			team.Team{Project: dir, Members: []team.Member{{Role: "cto"}, {Role: "qa"}}},
-			team.DefaultProfile,
-			"issue-96",
-			[]resumePlan{
-				{Role: "cto", Action: resumeLive, Note: "wake"},
-				{Role: "qa", Action: resumeLive, Note: "wake+presence"},
-			},
-			resumeExecOptions{Enabled: true, Terminal: "tmux", Target: "current-window", Layout: "vertical"},
-			false,
-		)
-	})
-	if err != nil {
-		t.Fatalf("all-live execResumePlan should succeed: %v", err)
-	}
-	for _, want := range []string{"resume --exec", "nothing to launch", "2 live"} {
-		if !strings.Contains(stdout, want) {
-			t.Errorf("output missing %q:\n%s", want, stdout)
-		}
-	}
-}
-
-func TestRunResumeTmuxPlanStagesLeadBeforeDependentsAcrossTargets(t *testing.T) {
-	for _, target := range []string{"current-window", "new-session", "new-window"} {
-		t.Run(target, func(t *testing.T) {
-			oldRun := runTmuxLaunchPlanForResume
-			oldVerify := verifyResumeExecLaunchRecordsNow
-			oldReady := verifyResumeLeadReadyNow
-			t.Cleanup(func() {
-				runTmuxLaunchPlanForResume = oldRun
-				verifyResumeExecLaunchRecordsNow = oldVerify
-				verifyResumeLeadReadyNow = oldReady
-			})
-
-			var events []string
-			var submitted []tmuxLaunchPlan
-			runTmuxLaunchPlanForResume = func(plan tmuxLaunchPlan) error {
-				events = append(events, "run:"+plan.Panes[0].Role)
-				submitted = append(submitted, plan)
-				return nil
-			}
-			verifyResumeExecLaunchRecordsNow = func(checks []resumeExecLaunchCheck, _ map[string]resumeExecLaunchSnapshot) []resumeExecLaunchResult {
-				events = append(events, "record:"+checks[0].Role)
-				out := make([]resumeExecLaunchResult, 0, len(checks))
-				for _, check := range checks {
-					out = append(out, resumeExecLaunchResult{Check: check, State: resumeExecLaunchStateLaunched})
-				}
-				return out
-			}
-			verifyResumeLeadReadyNow = func(check resumeExecLaunchCheck) error {
-				events = append(events, "ready:"+check.Role)
-				return nil
-			}
-
-			checks := []resumeExecLaunchCheck{{Role: "cto", Handle: "cto"}, {Role: "qa", Handle: "qa"}}
-			results, err := runResumeTmuxPlanWithLeadGate(
-				team.Team{Orchestrated: true, Lead: "cto"}, team.DefaultProfile, "issue-473",
-				tmuxLaunchPlan{Session: "squad", Workstream: "issue-473", Target: target, Layout: "tiled", Panes: []teamLaunchPane{
-					{Role: "cto", Command: "lead"}, {Role: "qa", Command: "worker"},
-				}}, checks, map[string]resumeExecLaunchSnapshot{}, false,
-			)
-			if err != nil {
-				t.Fatal(err)
-			}
-			if len(results) != 2 {
-				t.Fatalf("results = %+v", results)
-			}
-			if got := strings.Join(events, ","); got != "run:cto,record:cto,ready:cto,run:qa,record:qa" {
-				t.Fatalf("staging order = %s", got)
-			}
-			if len(submitted) != 2 || len(submitted[0].Panes) != 1 || len(submitted[1].Panes) != 1 || submitted[1].Panes[0].Role != "qa" {
-				t.Fatalf("submitted plans = %+v", submitted)
-			}
-			if !submitted[1].AllowExistingSession {
-				t.Fatalf("dependent %s plan must reuse the lead's terminal topology", target)
-			}
-		})
-	}
-}
-
-func TestRunResumeTmuxPlanLeadReadinessFailureLaunchesNoDependentsEvenWithForce(t *testing.T) {
-	oldRun := runTmuxLaunchPlanForResume
-	oldVerify := verifyResumeExecLaunchRecordsNow
-	oldReady := verifyResumeLeadReadyNow
-	t.Cleanup(func() {
-		runTmuxLaunchPlanForResume = oldRun
-		verifyResumeExecLaunchRecordsNow = oldVerify
-		verifyResumeLeadReadyNow = oldReady
-	})
-
-	var submitted []string
-	runTmuxLaunchPlanForResume = func(plan tmuxLaunchPlan) error {
-		for _, pane := range plan.Panes {
-			submitted = append(submitted, pane.Role)
-		}
-		return nil
-	}
-	verifyResumeExecLaunchRecordsNow = func(checks []resumeExecLaunchCheck, _ map[string]resumeExecLaunchSnapshot) []resumeExecLaunchResult {
-		return []resumeExecLaunchResult{{Check: checks[0], State: resumeExecLaunchStateLaunched}}
-	}
-	verifyResumeLeadReadyNow = func(resumeExecLaunchCheck) error { return stubErr("bootstrap mismatch") }
-
-	checks := []resumeExecLaunchCheck{{Role: "cto", Handle: "cto", Force: true}, {Role: "qa", Handle: "qa", Force: true}}
-	_, err := runResumeTmuxPlanWithLeadGate(
-		team.Team{Orchestrated: true, Lead: "cto"}, team.DefaultProfile, "issue-473",
-		tmuxLaunchPlan{Target: "current-window", Panes: []teamLaunchPane{{Role: "cto"}, {Role: "qa"}}},
-		checks, map[string]resumeExecLaunchSnapshot{}, false,
-	)
-	if err == nil || !strings.Contains(err.Error(), "lead readiness failed for cto") || !strings.Contains(err.Error(), "dependent roles were not launched") {
-		t.Fatalf("readiness error = %v", err)
-	}
-	if got := strings.Join(submitted, ","); got != "cto" {
-		t.Fatalf("submitted roles = %q, want lead only", got)
-	}
-}
-
-// #655: --skip-lead-check is the recovery escape hatch — a failing readiness
-// gate must not block dependent launches when the operator explicitly bypasses
-// it, and the bypass must never silently consult the gate.
-func TestRunResumeTmuxPlanSkipLeadCheckLaunchesDependentsDespiteFailingGate(t *testing.T) {
-	oldRun := runTmuxLaunchPlanForResume
-	oldVerify := verifyResumeExecLaunchRecordsNow
-	oldReady := verifyResumeLeadReadyNow
-	t.Cleanup(func() {
-		runTmuxLaunchPlanForResume = oldRun
-		verifyResumeExecLaunchRecordsNow = oldVerify
-		verifyResumeLeadReadyNow = oldReady
-	})
-
-	var submitted []string
-	runTmuxLaunchPlanForResume = func(plan tmuxLaunchPlan) error {
-		for _, pane := range plan.Panes {
-			submitted = append(submitted, pane.Role)
-		}
-		return nil
-	}
-	verifyResumeExecLaunchRecordsNow = func(checks []resumeExecLaunchCheck, _ map[string]resumeExecLaunchSnapshot) []resumeExecLaunchResult {
-		out := make([]resumeExecLaunchResult, 0, len(checks))
-		for _, check := range checks {
-			out = append(out, resumeExecLaunchResult{Check: check, State: resumeExecLaunchStateLaunched})
-		}
-		return out
-	}
-	gateConsulted := false
-	verifyResumeLeadReadyNow = func(resumeExecLaunchCheck) error {
-		gateConsulted = true
-		return stubErr("lead pane %217 is not live")
-	}
-
-	checks := []resumeExecLaunchCheck{{Role: "cto", Handle: "cto"}, {Role: "qa", Handle: "qa"}}
-	results, err := runResumeTmuxPlanWithLeadGate(
-		team.Team{Orchestrated: true, Lead: "cto"}, team.DefaultProfile, "issue-473",
-		tmuxLaunchPlan{Target: "current-window", Panes: []teamLaunchPane{{Role: "cto"}, {Role: "qa"}}},
-		checks, map[string]resumeExecLaunchSnapshot{}, true,
-	)
-	if err != nil {
-		t.Fatalf("skip-lead-check resume failed: %v", err)
-	}
-	if gateConsulted {
-		t.Fatal("--skip-lead-check must bypass the readiness gate entirely")
-	}
-	if got := strings.Join(submitted, ","); got != "cto,qa" {
-		t.Fatalf("submitted roles = %q, want lead then dependents", got)
-	}
-	if len(results) != 2 {
-		t.Fatalf("results = %+v", results)
-	}
-}
-
-func TestRunResumeTmuxPlanChecksAlreadyLiveLeadBeforePartialWorkerResume(t *testing.T) {
-	dir := t.TempDir()
-	setupFakeAMQSessionRoots(t)
-	resumeChdir(t, dir)
-	tm := team.Team{Project: dir, Orchestrated: true, Lead: "cto", Members: []team.Member{
-		{Role: "cto", Handle: "cto", Binary: "codex", Session: "issue-473"},
-		{Role: "qa", Handle: "qa", Binary: "codex", Session: "issue-473"},
-	}}
-
-	oldRun := runTmuxLaunchPlanForResume
-	oldVerify := verifyResumeExecLaunchRecordsNow
-	oldReady := verifyResumeLeadReadyNow
-	t.Cleanup(func() {
-		runTmuxLaunchPlanForResume = oldRun
-		verifyResumeExecLaunchRecordsNow = oldVerify
-		verifyResumeLeadReadyNow = oldReady
-	})
-	var events []string
-	verifyResumeLeadReadyNow = func(check resumeExecLaunchCheck) error {
-		events = append(events, "ready:"+check.Role)
-		return nil
-	}
-	runTmuxLaunchPlanForResume = func(plan tmuxLaunchPlan) error {
-		events = append(events, "run:"+plan.Panes[0].Role)
-		return nil
-	}
-	verifyResumeExecLaunchRecordsNow = func(checks []resumeExecLaunchCheck, _ map[string]resumeExecLaunchSnapshot) []resumeExecLaunchResult {
-		events = append(events, "record:"+checks[0].Role)
-		return []resumeExecLaunchResult{{Check: checks[0], State: resumeExecLaunchStateLaunched}}
-	}
-	workerChecks, err := buildResumeExecLaunchChecks(tm, []teamLaunchPane{{Role: "qa", CWD: dir}}, team.DefaultProfile, "issue-473", false)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if _, err := runResumeTmuxPlanWithLeadGate(tm, team.DefaultProfile, "issue-473", tmuxLaunchPlan{
-		Target: "new-window", Panes: []teamLaunchPane{{Role: "qa"}},
-	}, workerChecks, map[string]resumeExecLaunchSnapshot{}, false); err != nil {
-		t.Fatal(err)
-	}
-	if got := strings.Join(events, ","); got != "ready:cto,run:qa,record:qa" {
-		t.Fatalf("partial-resume order = %s", got)
-	}
-}
-
-func TestInspectResumeLeadReadyIgnoresLegacyBootstrapExpectation(t *testing.T) {
-	dir := t.TempDir()
-	agentDir := filepath.Join(dir, "agents", "cto")
-	root := dir
-	now := time.Now().UTC()
-	rec := launch.Record{
-		CWD: dir, Binary: "codex", Role: "cto", Handle: "cto", Session: "issue-473", Root: root,
-		AgentPID: 4242, StartedAt: now, TeamProfile: team.DefaultProfile,
-		BootstrapExpectation: &bootstrapack.Expectation{Required: true, LaunchID: "legacy-required"},
-		Tmux:                 &launch.TmuxInfo{PaneID: "%7", Session: "squad", Target: "new-window"},
-	}
-	if err := launch.Write(agentDir, rec); err != nil {
-		t.Fatal(err)
-	}
-	oldInspect := statusPaneInspector
-	statusPaneInspector = func(id string) (tmuxpane.TmuxPane, bool) {
-		return tmuxpane.TmuxPane{PaneID: id, Title: paneTitleToken("issue-473", "cto")}, id == "%7"
-	}
-	t.Cleanup(func() { statusPaneInspector = oldInspect })
-	probe := duplicateLaunchProbe{
-		PIDAlive: func(pid int) bool { return pid == 4242 },
-		ProcessMatch: func(_ int, predicate func(string) bool) bool {
-			return predicate("codex")
-		},
-		Now: func() time.Time { return now.Add(time.Second) },
-	}
-	check := resumeExecLaunchCheck{Role: "cto", Handle: "cto", Binary: "codex", CWD: dir, AgentDir: agentDir, Root: root, Workstream: "issue-473", Profile: team.DefaultProfile}
-	if ready, detail := inspectResumeLeadReady(check, probe); !ready || !strings.Contains(detail, "role cto live in pane %7") || strings.Contains(detail, "bootstrap") {
-		t.Fatalf("legacy bootstrap expectation affected readiness = %t, %q", ready, detail)
-	}
-}
-
-// #655: after an in-place lead restart (e.g. Claude Code /upgrade re-execs in
-// the same pane), the fresh process rewrites the pane's visible title. The
-// readiness gate must corroborate the recorded pane through the live agent's
-// controlling tty, report ready, and re-stamp the durable discovery token so
-// later checks pass on the primary path again.
-func TestInspectResumeLeadReadySurvivesInPlaceLeadRestart(t *testing.T) {
-	dir := t.TempDir()
-	agentDir := filepath.Join(dir, "agents", "lead")
-	root := dir
-	now := time.Now().UTC()
-	rec := launch.Record{
-		CWD: dir, Binary: "claude", Role: "lead", Handle: "lead", Session: "issue-655", Root: root,
-		AgentPID: 184, AgentTTY: "/dev/ttys011", StartedAt: now, TeamProfile: team.DefaultProfile,
-		Tmux: &launch.TmuxInfo{PaneID: "%217", Session: "loco", Target: "new-window"},
-	}
-	if err := launch.Write(agentDir, rec); err != nil {
-		t.Fatal(err)
-	}
-	oldInspect := statusPaneInspector
-	statusPaneInspector = func(id string) (tmuxpane.TmuxPane, bool) {
-		// The upgraded CLI set its own title; no @amq_squad_title option yet.
-		return tmuxpane.TmuxPane{PaneID: id, Title: "✳ upgraded CLI title"}, id == "%217"
-	}
-	oldTTY := statusPaneTTYInspector
-	paneTTY := "/dev/ttys011"
-	statusPaneTTYInspector = func(id string) (string, bool) {
-		return paneTTY, id == "%217"
-	}
-	oldRestamp := restampPaneDiscoveryToken
-	var restamped []string
-	restampPaneDiscoveryToken = func(paneID, workstream, role string) error {
-		restamped = append(restamped, paneID+"|"+workstream+"|"+role)
-		return nil
-	}
-	t.Cleanup(func() {
-		statusPaneInspector = oldInspect
-		statusPaneTTYInspector = oldTTY
-		restampPaneDiscoveryToken = oldRestamp
-	})
-	probe := duplicateLaunchProbe{
-		PIDAlive:     func(pid int) bool { return pid == 184 },
-		ProcessMatch: func(_ int, predicate func(string) bool) bool { return predicate("claude") },
-		ProcessTTY:   func(int) (string, bool) { return "/dev/ttys011", true },
-		Now:          func() time.Time { return now.Add(time.Second) },
-	}
-	check := resumeExecLaunchCheck{Role: "lead", Handle: "lead", Binary: "claude", CWD: dir, AgentDir: agentDir, Root: root, Workstream: "issue-655", Profile: team.DefaultProfile}
-	if ready, detail := inspectResumeLeadReady(check, probe); !ready || !strings.Contains(detail, "role lead live in pane %217") {
-		t.Fatalf("in-place restart readiness = %t, %q", ready, detail)
-	}
-	if want := []string{"%217|issue-655|lead"}; !reflect.DeepEqual(restamped, want) {
-		t.Fatalf("restamped = %v, want %v", restamped, want)
-	}
-
-	// A pane on a DIFFERENT pty is not the lead's pane (reused pane id after a
-	// tmux server restart): the gate must keep refusing.
-	restamped = nil
-	paneTTY = "/dev/ttys042"
-	if ready, detail := inspectResumeLeadReady(check, probe); ready || !strings.Contains(detail, "lead pane %217 is not live") {
-		t.Fatalf("mismatched pane tty readiness = %t, %q", ready, detail)
-	}
-	if len(restamped) != 0 {
-		t.Fatalf("refused pane must not be restamped: %v", restamped)
-	}
-}
-
-// TestExecResumePlanReportsPartialLaunchRecordFailure covers #208's
-// current-window failure mode: tmux accepted a multi-role plan, but one
-// requested role never published a fresh launch.json. The command must return
-// non-zero with role-level detail instead of leaving the operator with only the
-// optimistic "Added team panes" notice.
-func TestExecResumePlanReportsPartialLaunchRecordFailure(t *testing.T) {
-	dir := t.TempDir()
-	base := setupFakeAMQSessionRoots(t)
-
-	oldRun := runTmuxLaunchPlanForResume
-	oldTimeout := resumeExecLaunchVerifyTimeout
-	oldInterval := resumeExecLaunchVerifyInterval
-	oldBudget := resumeExecLaunchStartupBudget
-	runTmuxLaunchPlanForResume = func(plan tmuxLaunchPlan) error {
-		if plan.Target != "current-window" {
-			t.Errorf("target = %q, want current-window", plan.Target)
-		}
-		if len(plan.Panes) != 2 {
-			t.Errorf("panes = %d, want 2", len(plan.Panes))
-		}
-		writeMemberLaunchRecord(t, base, "issue-96", "cto", launch.Record{
-			CWD:       dir,
-			Binary:    "codex",
-			Role:      "cto",
-			StartedAt: time.Now().UTC(),
-			Tmux:      &launch.TmuxInfo{PaneID: "%101", Session: "squad", Target: "current-window"},
-		})
-		_, _ = os.Stderr.WriteString("Added 2 team pane(s) to current tmux window.\n")
-		return nil
-	}
-	resumeExecLaunchVerifyTimeout = time.Millisecond
-	resumeExecLaunchVerifyInterval = time.Millisecond
-	// The frontend-dev member never publishes a record, so without a short
-	// startup budget this fixture pays the real 30s boot wait (#688).
-	resumeExecLaunchStartupBudget = time.Millisecond
-	t.Cleanup(func() {
-		runTmuxLaunchPlanForResume = oldRun
-		resumeExecLaunchVerifyTimeout = oldTimeout
-		resumeExecLaunchVerifyInterval = oldInterval
-		resumeExecLaunchStartupBudget = oldBudget
-	})
-
-	stdout, stderr, err := captureOutput(t, func() error {
-		return execResumePlan(
-			team.Team{
-				Project: dir,
-				Members: []team.Member{
-					{Role: "cto", Binary: "codex", Handle: "cto", Session: "issue-96"},
-					{Role: "frontend-dev", Binary: "codex", Handle: "frontend-dev", Session: "issue-96"},
-				},
-			},
-			team.DefaultProfile,
-			"issue-96",
-			[]resumePlan{
-				{Role: "cto", Action: resumeFresh, Command: "amq-squad agent up codex --role cto"},
-				{Role: "frontend-dev", Action: resumeFresh, Command: "amq-squad agent up codex --role frontend-dev"},
-			},
-			resumeExecOptions{Enabled: true, Terminal: "tmux", Target: "current-window", Layout: "tiled"},
-			false,
-		)
-	})
-	if err == nil {
-		t.Fatal("partial launch record failure should return an error")
-	}
-	if _, ok := err.(*PartialError); !ok {
-		t.Fatalf("want *PartialError, got %T: %v", err, err)
-	}
-	for _, want := range []string{"Added 2 team pane", "partial launch failure", "frontend-dev", "missing", "launch record"} {
-		if !strings.Contains(stderr, want) && !strings.Contains(err.Error(), want) {
-			t.Errorf("missing %q in stderr/error\nstdout:\n%s\nstderr:\n%s\nerr:\n%v", want, stdout, stderr, err)
-		}
-	}
-}
-
-func TestVerifyResumeExecLaunchRecordsAdoptsStaleRecordByPaneTitle(t *testing.T) {
-	dir := t.TempDir()
-	base := setupFakeAMQSessionRoots(t)
-	oldStarted := time.Now().Add(-5 * time.Minute).UTC()
-	writeMemberLaunchRecord(t, base, "issue-96", "cto", launch.Record{
-		CWD:       dir,
-		Binary:    "codex",
-		Role:      "cto",
-		Handle:    "cto",
-		Session:   "issue-96",
-		StartedAt: oldStarted,
-		Tmux:      &launch.TmuxInfo{PaneID: "%old", Session: "squad", Target: "current-window"},
-	})
-	checks := []resumeExecLaunchCheck{{
-		Role:       "cto",
-		CWD:        dir,
-		AgentDir:   filepath.Join(base, "issue-96", "agents", "cto"),
-		Handle:     "cto",
-		Workstream: "issue-96",
-		Root:       filepath.Join(base, "issue-96"),
-		Binary:     "codex",
-		Profile:    team.DefaultProfile,
-	}}
-	snapshots := snapshotResumeExecLaunchRecords(checks)
-	withStubPaneLister(t, []tmuxpane.TmuxPane{{
-		Session:    "squad",
-		WindowID:   "@9",
-		WindowName: "issue-96",
-		PaneID:     "%77",
-		Title:      paneTitleToken("issue-96", "cto"),
-		Command:    "codex",
-		CWD:        dir,
-	}}, nil)
-	oldTimeout := resumeExecLaunchVerifyTimeout
-	oldInterval := resumeExecLaunchVerifyInterval
-	resumeExecLaunchVerifyTimeout = time.Millisecond
-	resumeExecLaunchVerifyInterval = time.Millisecond
-	t.Cleanup(func() {
-		resumeExecLaunchVerifyTimeout = oldTimeout
-		resumeExecLaunchVerifyInterval = oldInterval
-	})
-
-	results := verifyResumeExecLaunchRecords(checks, snapshots)
-	if len(results) != 1 || results[0].State != resumeExecLaunchStateLaunched {
-		t.Fatalf("verify results = %+v, want launched after pane adoption", results)
-	}
-	rec, err := launch.Read(checks[0].AgentDir)
-	if err != nil {
-		t.Fatalf("read adopted record: %v", err)
-	}
-	if rec.Tmux == nil || rec.Tmux.PaneID != "%77" || rec.Tmux.WindowID != "@9" || rec.Tmux.Target != "adopted" {
-		t.Fatalf("adopted tmux = %+v", rec.Tmux)
-	}
-	if !rec.StartedAt.After(oldStarted) {
-		t.Fatalf("StartedAt was not refreshed: got %s, old %s", rec.StartedAt, oldStarted)
-	}
-}
-
-// TestExecResumePlanRejectsUnknownTerminal makes sure the operator gets a
-// clear error rather than a downstream nil-map panic when the terminal
-// flag value is wrong.
-func TestExecResumePlanRejectsUnknownTerminal(t *testing.T) {
-	err := execResumePlan(
-		team.Team{Project: t.TempDir(), Members: []team.Member{{Role: "cto"}}},
-		team.DefaultProfile,
-		"issue-96",
-		[]resumePlan{{Role: "cto", Action: resumeRestore, Command: "echo hi"}},
-		resumeExecOptions{Enabled: true, Terminal: "screen", Target: "current-window", Layout: "vertical"},
-		false,
-	)
-	if err == nil || !strings.Contains(err.Error(), "unsupported terminal") {
-		t.Fatalf("expected unsupported-terminal error; got %v", err)
-	}
-}
+// gh#758/t11 slice B commit 3: TestExecResumePlanRefusesBlockedMembersUnlessForced,
+// TestExecResumePlanNothingToLaunch, TestRunResumeTmuxPlanStagesLeadBeforeDependentsAcrossTargets,
+// TestRunResumeTmuxPlanLeadReadinessFailureLaunchesNoDependentsEvenWithForce,
+// TestRunResumeTmuxPlanSkipLeadCheckLaunchesDependentsDespiteFailingGate,
+// TestRunResumeTmuxPlanChecksAlreadyLiveLeadBeforePartialWorkerResume,
+// TestInspectResumeLeadReadyIgnoresLegacyBootstrapExpectation,
+// TestInspectResumeLeadReadySurvivesInPlaceLeadRestart,
+// TestExecResumePlanReportsPartialLaunchRecordFailure,
+// TestVerifyResumeExecLaunchRecordsAdoptsStaleRecordByPaneTitle, and
+// TestExecResumePlanRejectsUnknownTerminal all deleted. Every one of them
+// exercised team_resume.go's own per-member launch orchestration internals
+// -- execResumePlan, resumePlan/resumeExecOptions/resumeBlocked/resumeLive/
+// resumeFresh/resumeRestore, runTmuxLaunchPlanForResume,
+// verifyResumeExecLaunchRecords(Now), inspectResumeLeadReady,
+// runResumeTmuxPlanWithLeadGate, buildResumeExecLaunchChecks,
+// snapshotResumeExecLaunchRecords -- all deleted along with the classifier,
+// no live equivalent under these names. --exec now drives
+// runResumeExec/buildSimpleStartPlan/runSimpleStartWithRequest (commit 2)
+// instead, which has its own, already-covered verification machinery
+// (verifySimpleStartRecords, validateCompleteTeamLaunchResult in
+// simple_start_test.go) and deliberately does NOT replicate the old
+// classifier's two-phase lead-then-dependents staged launch sequencing (a
+// launch-mechanics detail resume_exec.go's own doc comment calls out as
+// orthogonal to the new required-action lead gate, resolveResumeLeadGate,
+// which TestResumeExecRoleFilterDigestIgnoresLivenessOutsideSubset and the
+// team_lead_test.go lead-gate tests cover instead).
 
 // TestRunResumePositionalSessionHonored verifies that `resume <session>`
 // treats the positional as the session name, fixing #177's secondary finding.
@@ -1195,147 +647,14 @@ func walkDir(root string, fn func(path string, info os.FileInfo) error) error {
 	return nil
 }
 
-// The lead-main arrangement fires ONLY on the mid-run member-add signature:
-// orchestrated, current-window, lead already live (not in the plan), AND the
-// resume invoked from the lead's own recorded pane. Every other shape —
-// lead-only plan, lead being relaunched, a resume issued from an operator or
-// worker pane, or a --skip-lead-check bypass — clears the candidate flag
-// before the tmux plan runs.
-func TestRunResumeTmuxPlanLeadMainOnlyForDependentOnlyLaunch(t *testing.T) {
-	oldRun := runTmuxLaunchPlanForResume
-	oldVerify := verifyResumeExecLaunchRecordsNow
-	oldReady := verifyResumeLeadReadyNow
-	oldFromLead := resumeLaunchedFromLeadPaneNow
-	t.Cleanup(func() {
-		runTmuxLaunchPlanForResume = oldRun
-		verifyResumeExecLaunchRecordsNow = oldVerify
-		verifyResumeLeadReadyNow = oldReady
-		resumeLaunchedFromLeadPaneNow = oldFromLead
-	})
-	var flags []bool
-	runTmuxLaunchPlanForResume = func(plan tmuxLaunchPlan) error {
-		flags = append(flags, plan.LeadMainCurrentWindow)
-		return nil
-	}
-	verifyResumeExecLaunchRecordsNow = func(checks []resumeExecLaunchCheck, _ map[string]resumeExecLaunchSnapshot) []resumeExecLaunchResult {
-		out := make([]resumeExecLaunchResult, 0, len(checks))
-		for _, check := range checks {
-			out = append(out, resumeExecLaunchResult{Check: check, State: resumeExecLaunchStateLaunched})
-		}
-		return out
-	}
-	verifyResumeLeadReadyNow = func(resumeExecLaunchCheck) error { return nil }
-	launchedFromLeadPane := true
-	resumeLaunchedFromLeadPaneNow = func(resumeExecLaunchCheck) bool { return launchedFromLeadPane }
-
-	base := tmuxLaunchPlan{Target: "current-window", LeadMainCurrentWindow: true}
-	tm := team.Team{Orchestrated: true, Lead: "cto"}
-
-	// Mid-run add: dependents only, invoked from the lead pane — flag survives.
-	plan := base
-	plan.Panes = []teamLaunchPane{{Role: "researcher"}}
-	if _, err := runResumeTmuxPlanWithLeadGate(tm, team.DefaultProfile, "s", plan,
-		[]resumeExecLaunchCheck{{Role: "researcher", Handle: "researcher"}, {Role: "cto", Handle: "cto"}},
-		map[string]resumeExecLaunchSnapshot{}, false); err != nil {
-		t.Fatal(err)
-	}
-
-	// Lead-only plan: cleared.
-	plan = base
-	plan.Panes = []teamLaunchPane{{Role: "cto"}}
-	if _, err := runResumeTmuxPlanWithLeadGate(tm, team.DefaultProfile, "s", plan,
-		[]resumeExecLaunchCheck{{Role: "cto", Handle: "cto"}},
-		map[string]resumeExecLaunchSnapshot{}, false); err != nil {
-		t.Fatal(err)
-	}
-
-	// Lead + dependents relaunch: cleared on both the lead and dependent plans.
-	plan = base
-	plan.Panes = []teamLaunchPane{{Role: "cto"}, {Role: "qa"}}
-	if _, err := runResumeTmuxPlanWithLeadGate(tm, team.DefaultProfile, "s", plan,
-		[]resumeExecLaunchCheck{{Role: "cto", Handle: "cto"}, {Role: "qa", Handle: "qa"}},
-		map[string]resumeExecLaunchSnapshot{}, false); err != nil {
-		t.Fatal(err)
-	}
-
-	// Dependents only, but the resume is NOT running in the lead's recorded
-	// pane (operator shell, worker pane): the lead is live somewhere, yet
-	// arranging main-vertical here would rearrange the CALLER's window.
-	launchedFromLeadPane = false
-	plan = base
-	plan.Panes = []teamLaunchPane{{Role: "researcher"}}
-	if _, err := runResumeTmuxPlanWithLeadGate(tm, team.DefaultProfile, "s", plan,
-		[]resumeExecLaunchCheck{{Role: "researcher", Handle: "researcher"}, {Role: "cto", Handle: "cto"}},
-		map[string]resumeExecLaunchSnapshot{}, false); err != nil {
-		t.Fatal(err)
-	}
-
-	// --skip-lead-check: no live lead verified at all, so the flag must not
-	// survive even when the pane tie would have matched.
-	launchedFromLeadPane = true
-	plan = base
-	plan.Panes = []teamLaunchPane{{Role: "researcher"}}
-	if _, err := runResumeTmuxPlanWithLeadGate(tm, team.DefaultProfile, "s", plan,
-		[]resumeExecLaunchCheck{{Role: "researcher", Handle: "researcher"}, {Role: "cto", Handle: "cto"}},
-		map[string]resumeExecLaunchSnapshot{}, true); err != nil {
-		t.Fatal(err)
-	}
-
-	want := []bool{true, false, false, false, false, false}
-	if len(flags) != len(want) {
-		t.Fatalf("plan runs = %d, want %d (%v)", len(flags), len(want), flags)
-	}
-	for i, flag := range flags {
-		if flag != want[i] {
-			t.Fatalf("plan run %d LeadMainCurrentWindow = %t, want %t (%v)", i, flag, want[i], flags)
-		}
-	}
-}
-
-// resumeLaunchedFromLeadPane ties the arrangement to REAL identity evidence:
-// the lead launch record's pane id must equal the invoking $TMUX_PANE, and
-// every missing side fails closed.
-func TestResumeLaunchedFromLeadPaneComparesRecordedPane(t *testing.T) {
-	dir := t.TempDir()
-	writeRecord := func(paneID string) {
-		rec := launch.Record{Handle: "cto", Role: "cto", Tmux: &launch.TmuxInfo{PaneID: paneID}}
-		data, err := json.Marshal(rec)
-		if err != nil {
-			t.Fatal(err)
-		}
-		path := launch.Path(dir)
-		if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
-			t.Fatal(err)
-		}
-		if err := os.WriteFile(path, data, 0o644); err != nil {
-			t.Fatal(err)
-		}
-	}
-	check := resumeExecLaunchCheck{Role: "cto", Handle: "cto", AgentDir: dir}
-
-	writeRecord("%7")
-	t.Setenv("TMUX_PANE", "%7")
-	if !resumeLaunchedFromLeadPane(check) {
-		t.Fatal("matching recorded pane and TMUX_PANE not recognized")
-	}
-
-	t.Setenv("TMUX_PANE", "%9")
-	if resumeLaunchedFromLeadPane(check) {
-		t.Fatal("mismatched TMUX_PANE accepted as the lead pane")
-	}
-
-	t.Setenv("TMUX_PANE", "")
-	if resumeLaunchedFromLeadPane(check) {
-		t.Fatal("resume outside tmux accepted as the lead pane")
-	}
-
-	t.Setenv("TMUX_PANE", "%7")
-	writeRecord("")
-	if resumeLaunchedFromLeadPane(check) {
-		t.Fatal("record without a pane id accepted as the lead pane")
-	}
-
-	if resumeLaunchedFromLeadPane(resumeExecLaunchCheck{Role: "cto", Handle: "cto", AgentDir: filepath.Join(dir, "missing")}) {
-		t.Fatal("missing launch record accepted as the lead pane")
-	}
-}
+// gh#758/t11 slice B commit 3: TestRunResumeTmuxPlanLeadMainOnlyForDependentOnlyLaunch
+// and TestResumeLaunchedFromLeadPaneComparesRecordedPane deleted. Both tested
+// team_resume.go's own lead-main-current-window arrangement heuristic
+// (runResumeTmuxPlanWithLeadGate's LeadMainCurrentWindow flag,
+// resumeLaunchedFromLeadPane/resumeLaunchedFromLeadPaneNow) -- all deleted
+// with the classifier, no live equivalent. resume --exec's new shared-
+// machinery fold (commit 2) does not replicate this arrangement heuristic at
+// all: runSimpleStartWithRequest has no notion of "lead already live,
+// dependents launching from the lead's own pane" driving a different tmux
+// window layout: see resume_exec.go's own scope note on not replicating
+// "launch-mechanics detail orthogonal to" the new required-action gate.
