+++ internal/cli/runtime_json.go
@@ -1,7 +1,6 @@
 package cli
 
 import (
-	"io"
 	"os"
 	"path/filepath"
 	"sort"
@@ -14,7 +13,6 @@ import (
 	"github.com/omriariav/amq-squad/v2/internal/runtimecontrol"
 	"github.com/omriariav/amq-squad/v2/internal/team"
 	"github.com/omriariav/amq-squad/v2/internal/tmuxpane"
-	runwizard "github.com/omriariav/amq-squad/v2/internal/wizard"
 )
 
 // tmuxRuntimeJSON is the stable tmux runtime-identity block that amq-noc (and
@@ -453,177 +451,4 @@ func statusTopologyForRows(rows []statusRecord, orchestrated bool) *statusTopolo
 		topology.Detail = "tmux session topology is unknown"
 	}
 	return topology
-}
-
-// resumeMemberJSON is one member row in the resume_plan envelope. It mirrors the
-// human plan (role/action/note/command) and adds the runtime identity so a
-// client can decide whether to focus a live pane or re-open one.
-type resumeMemberJSON struct {
-	Role             string               `json:"role"`
-	Handle           string               `json:"handle,omitempty"`
-	Action           string               `json:"action"`
-	LaunchState      string               `json:"launch_state"`
-	RecordState      string               `json:"record_state"`
-	HasRestoreRecord bool                 `json:"has_restore_record"`
-	Wake             string               `json:"wake,omitempty"`
-	Note             string               `json:"note,omitempty"`
-	Command          string               `json:"command,omitempty"`
-	Tmux             *tmuxRuntimeJSON     `json:"tmux,omitempty"`
-	Terminal         *terminalRuntimeJSON `json:"terminal,omitempty"`
-	// Liveness is the shared liveness verdict (status + detail + signals), the
-	// SAME classification `status --json` reports. A client compares
-	// liveness.status to status's status instead of inferring liveness from the
-	// planning `action`. Omitted only on the blocked paths where no verdict ran.
-	Liveness *resumeLivenessJSON `json:"liveness,omitempty"`
-}
-
-// resumeLivenessJSON exposes the shared liveness verdict on a resume_plan member
-// so `resume --json` and `status --json` carry identical liveness for the same
-// agent. status is the same statusState string status emits.
-type resumeLivenessJSON struct {
-	Status  string        `json:"status"`
-	Detail  string        `json:"detail,omitempty"`
-	Signals statusSignals `json:"signals"`
-}
-
-// resumeEnvelopeData is the `resume_plan` envelope body: the same per-member
-// classification `resume` prints, in a stable shape clients can render as
-// actions. It is a read-only preview (never executes).
-type resumeEnvelopeData struct {
-	TeamHome                  string                     `json:"team_home"`
-	Workstream                string                     `json:"workstream"`
-	Profile                   string                     `json:"profile,omitempty"`
-	Mode                      string                     `json:"mode"`
-	NamespaceConflict         *namespaceConflictData     `json:"namespace_conflict,omitempty"`
-	Members                   int                        `json:"members"`
-	Plan                      []resumeMemberJSON         `json:"plan"`
-	GoalPlan                  *runwizard.ResumeGoalPlan  `json:"goal_plan,omitempty"`
-	NativeGoalBlockedRecovery []resumeNativeGoalRecovery `json:"native_goal_blocked_recovery,omitempty"`
-}
-
-// writeResumeJSON emits the resume_plan envelope. Pane liveness is resolved once
-// across every member that carries a tmux identity.
-func writeResumeJSON(out io.Writer, t team.Team, workstream string, mode resumeMode, profile string, conflict *namespaceConflictData, plans []resumePlan) error {
-	return writeResumeJSONWithGoal(out, t, workstream, mode, profile, conflict, plans, runwizard.ResumeGoalPlan{})
-}
-
-func writeResumeJSONWithGoal(out io.Writer, t team.Team, workstream string, mode resumeMode, profile string, conflict *namespaceConflictData, plans []resumePlan, goalPlan runwizard.ResumeGoalPlan) error {
-	rows := make([]resumeMemberJSON, 0, len(plans))
-	var livePanes map[string]bool
-	for _, p := range plans {
-		rt := tmuxRuntimeFromInfo(p.Tmux)
-		if rt != nil {
-			if livePanes == nil {
-				livePanes = livePaneIDSet(statusPaneLister)
-			}
-			fillPaneAliveFromLiveness(rt, livePanes, p.Liveness)
-		}
-		var liveness *resumeLivenessJSON
-		if p.Liveness != nil {
-			liveness = &resumeLivenessJSON{
-				Status:  string(p.Liveness.Status),
-				Detail:  p.Liveness.Detail,
-				Signals: p.Liveness.Signals,
-			}
-		}
-		var terminal *terminalRuntimeJSON
-		if p.RestoreRecord != nil {
-			terminal = terminalRuntimeFromInfo(p.RestoreRecord.Terminal)
-		}
-		if terminal == nil {
-			terminal = terminalRuntimeFromTmuxInfo(p.Tmux)
-		}
-		if terminal != nil {
-			if rt != nil && terminal.Backend == runtimecontrol.BackendTmux {
-				terminal.PaneAlive = rt.PaneAlive
-			}
-			if p.Liveness != nil {
-				terminal.PIDAlive = p.Liveness.RuntimeIdentity.PIDLive
-			}
-			row := statusRecord{Terminal: terminal}
-			if p.Liveness != nil {
-				row.Signals = p.Liveness.Signals
-			}
-			decorateTerminalRuntimeCapabilities(&row)
-		}
-		rows = append(rows, resumeMemberJSON{
-			Role:             p.Role,
-			Handle:           p.Handle,
-			Action:           string(p.Action),
-			LaunchState:      resumeLaunchState(p),
-			RecordState:      resumeRecordState(p),
-			HasRestoreRecord: p.HasRestoreRecord,
-			Wake:             wakeForJSON(p.Wake),
-			Note:             p.Note,
-			Command:          p.Command,
-			Tmux:             rt,
-			Terminal:         terminal,
-			Liveness:         liveness,
-		})
-	}
-	profile = strings.TrimSpace(profile)
-	if profile == team.DefaultProfile {
-		profile = ""
-	}
-	var goalPlanJSON *runwizard.ResumeGoalPlan
-	if goalPlan.SchemaVersion != 0 {
-		copy := goalPlan
-		goalPlanJSON = &copy
-	}
-	return writeJSONEnvelope(out, "resume_plan", resumeEnvelopeData{
-		TeamHome:                  canonicalFilesystemPath(t.Project),
-		Workstream:                workstream,
-		Profile:                   profile,
-		Mode:                      string(mode),
-		NamespaceConflict:         conflict,
-		Members:                   len(rows),
-		Plan:                      rows,
-		GoalPlan:                  goalPlanJSON,
-		NativeGoalBlockedRecovery: resumeNativeGoalBlockedRecoveries(plans),
-	})
-}
-
-func resumeLaunchState(p resumePlan) string {
-	switch {
-	case p.Action == resumeBlocked:
-		return "blocked"
-	case strings.TrimSpace(p.Command) != "":
-		return "will-launch"
-	case p.Action == resumeLive:
-		return "skipped"
-	default:
-		return "failed"
-	}
-}
-
-func resumeRecordState(p resumePlan) string {
-	if p.Liveness != nil {
-		switch p.Liveness.Status {
-		case statusStateLive, statusStateWakeLive:
-			if !p.Liveness.LaunchFound {
-				return "missing"
-			}
-			return "launched"
-		case statusStateStale:
-			if p.Liveness.LaunchFound {
-				return "stale-record"
-			}
-			return "stale-signal"
-		case statusStateMissing:
-			return "missing"
-		}
-	}
-	if p.HasRestoreRecord {
-		return "restorable"
-	}
-	return "missing"
-}
-
-// wakeForJSON normalizes the human "-" placeholder to an empty (omitted) field.
-func wakeForJSON(w string) string {
-	w = strings.TrimSpace(w)
-	if w == "-" {
-		return ""
-	}
-	return w
 }
